crud / search working

+++ game/views.py
@@ -1,45 +1,58 @@
 from django.shortcuts import render, redirect
-from django.http import JsonResponse
 from .models import Game
 from .forms import GameForm
+from .filter import GameFilter
 
 
-def index(request):
-    games = Game.objects.all()
-    return render(request, 'games/home.html', {'games': games})
+def game_index(request):
+    if 'f' in request.GET:
+        f = request.GET['f']
+        games = Game.objects.filter(title__icontains=f)
+    else:
+        games = Game.objects.all()
+    gamefilter = GameFilter()
+    context = {'games': games, 'gamefilter':gamefilter}
+    return render(request, 'games/home.html', context)
 
 def game_list(request):
     games = Game.objects.all()
-    return render(request, 'games/game_list.html', {'games': games})
+    context = {'games': games}
+    return render(request, 'games/game_list.html', context)
 
 def game_detail(request, pk):
-    game = Game.objects.get(pk=pk)
-    return render(request, 'games/game_detail.html', {'game': game})
+    games = Game.objects.get(pk=pk)
+    context = {'games': games}
+    return render(request, 'games/game_detail.html', context)
 
-def game_create(request):
-    form = GameForm(request.POST or None)
+def add_game(request):
+    form = GameForm()
     if request.method == 'POST':
+        #print('Printing POST:', request.POST)
         form = GameForm(request.POST)
         if form.is_valid():
-            game_create = form.save()
-            return redirect('index')
-    else:
-        form = GameForm()
-    return render(request, 'games/game_form.html', {'form': form})
+            form.save()
+            return redirect('/')
+        
+    context = {'form': form}
+    return render(request, 'games/add_game.html', context)
 
-def game_update(request, pk):
-    game = Game.objects.get(pk=pk)
+def edit_game(request, pk):
+    game = Game.objects.get(id=pk)
+    form = GameForm(instance=game)
     if request.method == 'POST':
+        #print('Printing POST:', request.POST)
         form = GameForm(request.POST, instance=game)
         if form.is_valid():
             form.save()
-            return redirect('game_list')
-    else:
-        form = GameForm(instance=game)
-    return render(request, 'games/game_form.html', {'form': form})
+            return redirect('/')
+        
+    context = {'form': form}
+    return render(request, 'games/add_game.html', context)
 
-def game_delete(request, pk):
+def delete_game(request, pk):
     game = Game.objects.get(id=pk)
-    game.delete()
-    return render('index')
+    if request.method == 'POST':
+        game.delete()
+        return redirect('/')
+    return render(request, 'games/game_confirm_delete.html', {'game': game})
